Release 1.1b - ContextualAxis micro-optimization
Micro-Optimiaztion specific to ContextualAxis.

As the vector2 created doesn't need to be kept as-is, Vector2.Normalize() is now used.

Vector2.normalized returns a new Vector2 object, which can bad for memory use if a new object isn't needed. Vector2.Normalize() modifies the current object instead of creating a new one. As the raw input isn't needed, this should be better for memory usage.